+++ tests/thread_jetbench/jetbench.h
@@ -9,7 +9,7 @@
 uint8_t benchmark(uint8_t enginetype);
 double getGama(double temp);
 double getCp(double temp);
-double getMach(uint32_t sub, double corair, double gama1);
+double getMach(uint8_t sub, double corair, double gama1);
 double getAir(double mach, double gama2);
 double getRayleighLoss(double mach1, double ttrat, double tlow);
 void defaultParam(void);
@@ -23,5 +23,5 @@ double fabs(double x);
 double log(double x);
 double expo(double x);
 double fpow(double x, double y);
-double power(double x, int32_t y);
+double power(double x, int8_t y);
 void* parallelKernel(void *args);
